requirements.txt
fastapi
uvicorn
openai-whisper
numpy==1.26.4
